added and configured typescript
