+++ src/TypesAndInterfaces/AppTypes.ts
@@ -1,0 +1,1 @@
+export type React$Node = JSX.Element | null;
